Rename modules to snake_case, add public API, create test stub dirs
Renames 9 modules from camelCase/concatenated to PEP 8 snake_case:
  hexfunctions      → hex_functions
  camanager         → ca_manager
  derobjectidentifier → der_object_identifier
  tlvparser         → tlv_parser
  attacks/aaTraceability    → attacks/aa_traceability
  attacks/bruteForce        → attacks/brute_force
  attacks/errorFingerprinting → attacks/error_fingerprinting
  attacks/macTraceability   → attacks/mac_traceability
  attacks/signEverything    → attacks/sign_everything

Updates all import references across the codebase to match new names.
Adds EPassport, EPassportException, MRZ to package __init__.py for
direct top-level imports. Creates ePassportViewer/tests/ stub dir.

Incorporates change from claude/file-structure-review-NEW2h.
