+++ pypassport/src/pypassport/__init__.py
@@ -1,0 +1,2 @@
+from pypassport.epassport import EPassport, EPassportException
+from pypassport.doc9303.mrz import MRZ
